Tela para pesquisar os tipos de propaganda.

+++ PPVR.WebApp/ViewModels/TipoPropaganda/TipoPropagandaViewModel.cs
@@ -6,7 +6,7 @@ namespace PPVR.WebApp.ViewModels.TipoPropaganda
     public class TipoPropagandaViewModel
     {
         [Key]
-        public short TipoPropagandaId { get; set; }
+        public int TipoPropagandaId { get; set; }
 
         [Display(Name = nameof(Labels.TipoPropagandaDescricao), ResourceType = typeof(Labels))]
         [Required(ErrorMessageResourceType = typeof(ValidationErrorMessage),
